create macOS-specific format-commit.sh
The macOS-default xargs, sed and grep are far out of step with the GNU versions. The simplest way to make this work on macOS is to install the GNU versions and use a separate script. By default homebrew prepends a 'g' to each command, so this example accounts for that. To get the GNU versions, do 'brew install gnu-sed grep findutils'.